Added in an example file.
This contains D5 (d=1,2), D6 (d=1,2), E7(d=1,2), E8(d=1,2), RNC5, RNC6.
No assertions are yet in, and thetas and syzygy information are not added in yet.  The code David+Mike were using
is added at the beginning of examples2.m2.  It doesn't do as much as Mahrud's code,
but it should give us 2 pieces of code to check consistency.